Fixed accounts sorting incorrectly
