+++ .nvmrc
@@ -1,1 +1,1 @@
-v16.13.0
+v16.15.1
