+++ src/pages/popup/components/tab-contents/TabRanking.tsx
@@ -11,6 +11,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import useGgbb528Accounts from '../../hooks/useGgbb528Accounts';
 import LiveIcon from '../LiveIcon';
 import useMultipleOPGGSpectates from '../../hooks/useMultipleOPGGSpectates';
+import useAccountLiveHistory from '../../hooks/useAccountLiveHistory';
 
 function ServerOptionButtons({
   account,
@@ -25,6 +26,27 @@ function ServerOptionButtons({
   const { data: accounts } = useGgbb528Accounts();
   const maxIndex = accounts?.length || 0;
   const spectates = useMultipleOPGGSpectates();
+  const { histories } = useAccountLiveHistory();
+
+  const sortFn = (a: Account, b: Account) => {
+    const aHistory = histories?.find((h) => h.summoner_id === a.summoner_id);
+    const bHistory = histories?.find((h) => h.summoner_id === b.summoner_id);
+
+    if (aHistory && bHistory) {
+      const aLastGameTime = new Date(aHistory.last_game_start_time);
+      const bLastGameTime = new Date(bHistory.last_game_start_time);
+
+      return bLastGameTime.getTime() - aLastGameTime.getTime();
+    }
+
+    if (aHistory) return -1;
+    if (bHistory) return 1;
+    return 0;
+  };
+
+  useEffect(() => {
+    accounts?.sort(sortFn);
+  }, [accounts, histories]);
 
   const [liveSpectates] = spectates.filter(
     (spectate) => spectate.status === 'success'
@@ -91,7 +113,7 @@ function ServerOptionButtons({
                   </Pill>
                   <span>{accountInfo.account_id}</span>
                   {liveSpectates?.data?.account.summonerId ===
-                    accountInfo.summoner_id && <LiveIcon />}
+                    accountInfo.summoner_id && <LiveIcon showTooltip={false} />}
                 </div>
               </a>
             </li>
